qmd engine

+++ R/render.R
@@ -6,6 +6,8 @@
 #' @param header_latex optional .tex file of header latex
 #' @param keep_tex Boolean indicating if to keep resulting .tex file from latex conversion. Defaults to FALSE.
 #' @param escape_latex Boolean indicating if headers and footers of a gt table should be escaped with gt::escape_latex
+#' @param quarto Boolean indicating whether to use Quarto as the rendering
+#'   engine. Defaults to `FALSE`, which uses Rmarkdown to render. `r lifecycle::badge("experimental")`
 #'
 #' @returns This function saves a pdf to a specified location
 #' @export
@@ -24,7 +26,13 @@
 #'  render_pdf()
 #' ```
 #'
-render_pdf <- function(x, display_loc = NULL, transform = NULL, header_latex = NULL, keep_tex = FALSE, escape_latex = TRUE){
+render_pdf <- function(x,
+                       display_loc = NULL,
+                       transform = NULL,
+                       header_latex = NULL,
+                       keep_tex = FALSE,
+                       escape_latex = TRUE,
+                       quarto = FALSE){
 
   if (!inherits(x, "docorator")) {
     cli::cli_abort("The {.arg {rlang::caller_arg(x)}} argument must be class docorator, not {.obj_type_friendly {x}}. See documentation for `as_docorator`.",
@@ -63,40 +71,47 @@ render_pdf <- function(x, display_loc = NULL, transform = NULL, header_latex = N
 
   # set name if needed
   if (is.null(x$display_name)){
-    filename <- "doc.pdf"
-  }else{
-    filename <- paste0(x$display_name,".pdf")
+    x$display_name <- "docorator"
   }
+  filename <- paste0(x$display_name,".pdf")
 
-  # render pdf
-  doc <- rmarkdown::render(file.path(temp_dir, "template", "template.Rmd"),
-                output_file = filename,
-                output_dir = display_loc,
-                output_options = list(keep_tex = keep_tex),
-                params = list(x = x$display,
-                              header = hf_process(x$header, escape_latex = escape_latex),
-                              footer = hf_process(x$footer, escape_latex = escape_latex),
-                              geometry = geom_process(
-                                x$header,
-                                x$footer,
-                                x$fontsize,
-                                x$geometry
-                              ),
-                              fontsize = x$fontsize,
-                              fig_dim = x$fig_dim,
-                              transform = transform
-                ),
-                quiet = TRUE)
+  if(quarto){
+    withr::with_envvar(
+      new = c("DOCORATOR_RENDER_ENGINE" = "qmd"),
+      render_pdf_qmd(x, display_loc, transform, header_latex, clean=!keep_tex)
+    )
+  } else {
 
-  if (!is.null(doc)){
-    cli::cli_alert_success("Document created at: {doc}")
+    # render rmd -> pdf
+    doc <- rmarkdown::render(file.path(temp_dir, "template", "template.Rmd"),
+                             output_file = filename,
+                             output_dir = display_loc,
+                             output_options = list(keep_tex = keep_tex),
+                             params = list(x = x$display,
+                                           header = hf_process(x$header, escape_latex = escape_latex),
+                                           footer = hf_process(x$footer, escape_latex = escape_latex),
+                                           geometry = geom_process(
+                                             x$header,
+                                             x$footer,
+                                             x$fontsize,
+                                             x$geometry
+                                           ),
+                                           fontsize = x$fontsize,
+                                           fig_dim = x$fig_dim,
+                                           transform = transform
+                             ),
+                             quiet = TRUE)
+
+    if (!is.null(doc)){
+      cli_alert_success("Document created at: {doc}")
+    }
+
+    # clean up temp files
+    unlink(file.path(temp_dir, "template"), recursive = TRUE)
+
+    # return docorator object for further renders
+    invisible(x)
   }
-
-  # clean up temp files
-  unlink(file.path(temp_dir, "template"), recursive = TRUE)
-
-  # return docorator object for further renders
-  invisible(x)
 }
 
 
@@ -190,4 +205,133 @@ render_rtf <- function(x, display_loc = NULL, remove_unicode_ws = TRUE, use_page
   # return docorator object for further renders
   invisible(x)
 
+}
+
+
+#' Render to pdf (quarto)
+#'
+#' @param x `docorator` object
+#' @param display_loc path to save the output pdf to
+#' @param transform optional latex transformation function to apply to a gt latex string
+#' @param header_latex optional .tex file of header latex
+#' @param clean whether to clean the temporary docorator files. Defaults to `TRUE`
+#'
+#' @returns This function saves a pdf to a specified location
+#' @noRd
+#'
+#' @importFrom rmarkdown render
+#' @importFrom cli cli_alert_success cli_warn
+#' @importFrom tools file_ext
+#'
+#' @section Examples:
+#'
+#' ```r
+#' gt::gtcars |>
+#'   dplyr::slice_head(n = 10) |>
+#'   dplyr::select(mfr, model, year, msrp) |>
+#'   gt::gt(groupname_col = "mfr",
+#'          row_group_as_column = TRUE) |>
+#'   as_docorator(
+#'    header = fancyhead(fancyrow("Header 1"), fancyrow("Header 2")),
+#'    display_name = "mytbl") |>
+#'  render_pdf()
+#' ```
+#'
+render_pdf_qmd <- function(x,
+                           display_loc = NULL,
+                           transform = NULL,
+                           header_latex = NULL,
+                           clean = TRUE){
+
+  if (!inherits(x, "docorator")) {
+    cli_abort("The {.arg {caller_arg(x)}} argument must be class docorator, not {.obj_type_friendly {x}}. See documentation for `as_docorator`.",
+              call = caller_env())
+  }
+
+  # check transform is a function if not convert to NULL
+  if (!is.null(transform) & !inherits(transform, "function")) {
+    cli_warn("The transform argument must be a function, not {.obj_type_friendly {transform}}. No transform applied.",
+             call = caller_env())
+    transform <- NULL
+  }
+
+  # if no path is given, use docorator path
+  if(is.null(display_loc)){
+    display_loc <- x$display_loc
+  }
+
+  # set name if needed
+  if (is.null(x$display_name)){
+    x$display_name <- "docorator"
+  }
+  qmd_name <- paste0(x$display_name,".qmd")
+  pdf_name <- paste0(x$display_name, ".pdf")
+  docorator_name <- paste0(x$display_name, "_docorator_obj.Rds")
+
+  cur_dir <- getwd()
+  render_dir <- file.path(display_loc, paste0(x$display_name, "_docorator_files"))
+  if(!dir.exists(render_dir)){
+    dir.create(render_dir)
+  }
+
+  on.exit({
+    if(isTRUE(clean)){
+      unlink(render_dir, recursive = TRUE)
+    }
+  },
+  add = TRUE)
+
+  withr::with_dir(
+    new = render_dir,
+    code = {
+
+      # copy template qmd to a temp directory
+      template <- system.file("template", "template.qmd", package = "docorator")
+      file.copy(template, getwd(), overwrite = TRUE, recursive = TRUE)
+      file.rename("template.qmd", qmd_name)
+
+      # copy tex header to temp directory and rename if one exists
+      if(!is.null(header_latex)){
+        if(file.exists(header_latex) && file_ext(header_latex) == "tex"){
+          file_name <- basename(header_latex)
+          file.copy(header_latex, getwd(), overwrite = TRUE, recursive = TRUE)
+          file.rename(file_name, "header.tex")
+        }else{
+          cli_warn("The header_latex argument must point to a valid .tex file. No header options applied.",
+                   call = caller_env())
+        }
+      }
+
+      # save docorator obj to temp dir
+      saveRDS(x, docorator_name)
+
+      # render pdf
+      doc <- purrr::safely(
+        quarto::quarto_render)(
+          input = qmd_name,
+          output_format = "pdf",
+          output_file = pdf_name,
+          execute_params = list(
+            display_name = x$display_name,
+            pkg_path = ifelse(testthat::is_testing(),cur_dir,"")
+          ),
+          quiet = FALSE)
+
+      if (is.null(doc$error) && file.exists(pdf_name)){
+
+        out_path <- file.path(cur_dir, display_loc, pdf_name)
+        file_ok <- file.copy(from = pdf_name,
+                             to = out_path,
+                             overwrite = TRUE)
+
+        if (file_ok){
+          cli_alert_success("Document created at: {out_path}")
+        }
+      }
+
+    }
+  )
+
+  # return docorator object for further renders
+  invisible(x)
 }
